src/pydantic-ai-query-dataset/pyproject.toml
[project]
name = "pydantic-ai-query-dataset"
version = "0.1.0"
description = "Agent runs DuckDB SQL over a pandas DataFrame loaded from a remote JSON URL"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.1.0",
    "httpx>=0.27",
    "pandas>=2.2.0",
    "pydantic-ai>=1.7.0",
    "python-dotenv>=1.1.0",
]
